+++ app/Http/Controllers/Backend/VendorProductVariantItemController.php
@@ -7,23 +7,41 @@ use Illuminate\Http\Request;
 use App\Models\ProductVariant;
 use App\Models\ProductVariantItem;
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Auth;
 use App\DataTables\VendorProductVariantItemDataTable;
 
 class VendorProductVariantItemController extends Controller
 {
-    public function index($productId,$variantId,VendorProductVariantItemDataTable $dataTable){
+    public function index($productId, $variantId, VendorProductVariantItemDataTable $dataTable)
+    {
         $product = Product::findOrFail($productId);
         $variant = ProductVariant::findOrFail($variantId);
-        return $dataTable->render('vendor.product.product-variant-item.index' , compact('product','variant'));
+        /** check product owner */
+        if ($product->vendor_id !== Auth::user()->vendor->id) {
+            abort(404);
+        }
+        if ($variant->product->vendor_id !== Auth::user()->vendor->id) {
+            abort(404);
+        }
+        return $dataTable->render('vendor.product.product-variant-item.index', compact('product', 'variant'));
     }
 
-    public function create($productId,$variantId){
+    public function create($productId, $variantId)
+    {
         $product = Product::findOrFail($productId);
         $variant = ProductVariant::findOrFail($variantId);
-        return view('vendor.product.product-variant-item.create' , compact('product' , 'variant'));
+        /** check product owner */
+        if ($product->vendor_id !== Auth::user()->vendor->id) {
+            abort(404);
+        }
+        if ($variant->product->vendor_id !== Auth::user()->vendor->id) {
+            abort(404);
+        }
+        return view('vendor.product.product-variant-item.create', compact('product', 'variant'));
     }
 
-    public function store(Request $request){
+    public function store(Request $request)
+    {
 
         $request->validate([
             'variant_id' => 'required|integer',
@@ -41,18 +59,26 @@ class VendorProductVariantItemController extends Controller
         $variantItem->status = $request->status;
         $variantItem->save();
 
-        toastr('Created Successfully' , 'success');
-        return redirect()->route('vendor.variant-item.index' , ['productId' => $request->product_id ,'variantId' => $request->variant_id]);
-
+        toastr('Created Successfully', 'success');
+        return redirect()->route('vendor.variant-item.index', ['productId' => $request->product_id, 'variantId' => $request->variant_id]);
     }
 
-    public function edit(string $productId,string $variantItemId) {
+    public function edit(string $productId, string $variantItemId)
+    {
         $variantItem = ProductVariantItem::findOrFail($variantItemId);
         $product = Product::findOrFail($productId);
-        return view('vendor.product.product-variant-item.edit' , compact('variantItem' , 'product'));
+        /** check product owner */
+        if ($product->vendor_id !== Auth::user()->vendor->id) {
+            abort(404);
+        }
+        if($variantItem->productVariant->product->vendor_id !== Auth::user()->vendor->id){
+            abort(404);
+        }
+        return view('vendor.product.product-variant-item.edit', compact('variantItem', 'product'));
     }
 
-    public function update(Request $request , $variantItemId){
+    public function update(Request $request, $variantItemId)
+    {
         $request->validate([
             'name' => 'required|max:200',
             'price' => 'required|integer',
@@ -61,23 +87,31 @@ class VendorProductVariantItemController extends Controller
         ]);
 
         $variantItem = ProductVariantItem::findOrFail($variantItemId);
+        if($variantItem->productVariant->product->vendor_id !== Auth::user()->vendor->id){
+            abort(404);
+        }
         $variantItem->name = $request->name;
         $variantItem->price = $request->price;
         $variantItem->is_default = $request->is_default;
         $variantItem->status = $request->status;
         $variantItem->save();
 
-        toastr('Updated Successfully' , 'success');
-        return redirect()->route('vendor.variant-item.index' , ['productId' => $request->product_id ,'variantId' => $variantItem->product_variant_id]);
+        toastr('Updated Successfully', 'success');
+        return redirect()->route('vendor.variant-item.index', ['productId' => $request->product_id, 'variantId' => $variantItem->product_variant_id]);
     }
 
-    public function destroy(string $variantItemId){
+    public function destroy(string $variantItemId)
+    {
         $variantItem = ProductVariantItem::findOrfail($variantItemId);
+        if($variantItem->productVariant->product->vendor_id !== Auth::user()->vendor->id){
+            abort(404);
+        }
         $variantItem->delete();
-        return response(['status' => 'success' , 'message' => 'Deleted Successfully']);
+        return response(['status' => 'success', 'message' => 'Deleted Successfully']);
     }
 
-    public function changeStatus(Request $request){
+    public function changeStatus(Request $request)
+    {
         $variantItem = ProductVariantItem::findOrFail($request->id);
         $variantItem->status = $request->status == 'true' ? 1 : 0;
         $variantItem->save();
